[docs-infra] Move built-in type checks to their own module
