+++ packages/docs-infra/src/pipeline/loadServerTypesMeta/externalTypes.ts
@@ -9,6 +9,7 @@ import {
   isTypeOperatorType,
   isTypeQueryType,
 } from './typeGuards';
+import { isBuiltInTypeName } from './builtInTypes';
 import { groupType, UNION_OR_INTERSECTION } from './precedence';
 
 /**
@@ -37,63 +38,6 @@ export interface ExternalTypesCollector {
   pattern?: RegExp;
   /** Map of original export names to dotted display names, used to identify renamed own types */
   typeNameMap?: Record<string, string>;
-}
-
-/**
- * Built-in type namespaces that should not be collected as external types.
- */
-const BUILT_IN_NAMESPACES = ['React', 'JSX', 'HTML', 'CSS', 'SVG', 'Omit', 'Pick', 'Partial'];
-
-/**
- * Checks if a type name belongs to a built-in namespace that should be skipped
- * during external type collection.
- */
-export function isBuiltInTypeName(typeName: tae.TypeName): boolean {
-  const name = typeName.name || '';
-  return BUILT_IN_NAMESPACES.some(
-    (ns) => name.startsWith(ns) || (typeName.namespaces?.includes(ns) ?? false),
-  );
-}
-
-/** The name a type is written as, for the node kinds that carry one. */
-function namedAs(type: tae.AnyType): tae.TypeName | undefined {
-  return 'typeName' in type ? type.typeName : undefined;
-}
-
-/**
- * Whether a name is built-in, matched by whole segment rather than by prefix.
- *
- * `isBuiltInTypeName` matches prefixes, which is harmless where it decides whether to
- * repeat a definition. Here it decides whether a type's members are shown at all, so a
- * `PickerConfig` of ours must not be taken for the built-in `Pick`.
- */
-function namesBuiltIn(typeName: tae.TypeName): boolean {
-  return BUILT_IN_NAMESPACES.some(
-    (ns) => typeName.name === ns || (typeName.namespaces?.includes(ns) ?? false),
-  );
-}
-
-/** Whether a type names something this page is responsible for documenting. */
-function referencesDocumentedType(type: tae.AnyType): boolean {
-  const typeName = namedAs(type);
-  if (!typeName) {
-    // Literals and intrinsics name nothing, so they leave the verdict to their siblings.
-    return false;
-  }
-  if (!namesBuiltIn(typeName)) {
-    return true;
-  }
-  return (typeName.typeArguments ?? []).some((argument) => referencesDocumentedType(argument.type));
-}
-
-/**
- * Whether a type is one the reader already knows, rather than one this page documents.
- *
- * A utility wrapper qualifies only when everything it wraps is built-in too: the keys of
- * `Omit<Config, 'size'>` come from `Config`, so naming the wrapper would document nothing.
- */
-export function isBuiltInTypeReference(type: tae.AnyType): boolean {
-  return namedAs(type) !== undefined && !referencesDocumentedType(type);
 }
 
 /**
